requirements.txt
streamlit
pandas
openpyxl
google-generativeai
python-dotenv
fastapi
uvicorn[standard]
python-multipart
twilio
assemblyai
watchfiles
